backend/ai-automation-service/main.py
# main.py - The Core Autonomous Agent Service Engine
# This file contains the "engine" of Eliza's AI operations.

import asyncio
import logging
import os
import sys
import time
from datetime import datetime
from typing import Dict, Any

from web3 import Web3
import openai
from dotenv import load_dotenv

# --- Setup ---
sys.path.append(os.path.join(os.path.dirname(__file__), 'src'))
load_dotenv()

from agents.governance_agent import GovernanceAgent
from agents.treasury_agent import TreasuryAgent
from agents.community_agent import CommunityAgent
from agents.self_improvement_agent import SelfImprovementAgent
from utils.blockchain_utils import BlockchainUtils
from utils.ai_utils import AIUtils
from utils.github_utils import GitHubUtils
from utils.terminal_utils import TerminalUtils
from utils.browser_utils import BrowserUtils

# --- Logging and API Key Configuration ---
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', handlers=[logging.StreamHandler()])
logger = logging.getLogger(__name__)

# --- FIX for OpenAI Key ---
openai.api_key = os.getenv("OPENAI_API_KEY")
if not openai.api_key:
    logger.error("CRITICAL: OPENAI_API_KEY environment variable not found! AI features will fail.")

class AIAutomationService:
    """
    This class orchestrates all the autonomous AI agents.
    It's started and managed by launcher.py.
    """
    def __init__(self):
        self.blockchain_utils = BlockchainUtils()
        self.ai_utils = AIUtils()
        self.governance_agent = GovernanceAgent(self.blockchain_utils, self.ai_utils)
        self.treasury_agent = TreasuryAgent(self.blockchain_utils, self.ai_utils)
        self.community_agent = CommunityAgent(self.blockchain_utils, self.ai_utils)
        self.running = False
        logger.info("AIAutomationService engine initialized and ready.")

    async def start_automation(self):
        """Starts the main automation loop using pure asyncio."""
        logger.info("🚀 AI Engine Starting: Beginning main automation loop...")
        self.running = True

        while self.running:
            try:
                logger.info("--- Starting new agent cycle ---")
                # This runs all your agent tasks concurrently, which is highly efficient.
                await asyncio.gather(
                    self.governance_agent.run_cycle(), # You must create this method in your agent
                    self.treasury_agent.run_cycle(),   # You must create this method in your agent
                    self.community_agent.run_cycle(),  # You must create this method in your agent
                    self.self_improvement_agent.run_cycle()  # Autonomous self-improvement
                )
                logger.info("--- Agent cycle complete. Sleeping for 60 seconds. ---")
                await asyncio.sleep(60)

            except Exception as e:
                logger.error(f"FATAL ERROR in main agent cycle: {e}", exc_info=True)
                await asyncio.sleep(120)

    # ... (Your other methods like stop_automation, execute_manual_action, etc. go here)
