backend/ai-automation-service/src/utils/browser_utils.py
"""
Browser utilities for web research and documentation access
"""

import asyncio
import aiohttp
from bs4 import BeautifulSoup
import logging

logger = logging.getLogger(__name__)

class BrowserUtils:
    def __init__(self):
        self.session = None
    
    async def search(self, query):
        """Search for information online"""
        try:
            # Simple search implementation
            search_url = f"https://www.google.com/search?q={query.replace(' ', '+')}"
            
            async with aiohttp.ClientSession() as session:
                async with session.get(search_url) as response:
                    if response.status == 200:
                        content = await response.text()
                        # Parse search results (simplified)
                        return [search_url]  # Return URLs
                    
        except Exception as e:
            logger.error(f"[Browser] Search error: {e}")
            return []
    
    async def get_content(self, urls):
        """Get content from URLs"""
        try:
            content_data = []
            
            async with aiohttp.ClientSession() as session:
                for url in urls[:3]:  # Limit to first 3 URLs
                    try:
                        async with session.get(url) as response:
                            if response.status == 200:
                                html = await response.text()
                                soup = BeautifulSoup(html, 'html.parser')
                                text = soup.get_text()
                                content_data.append({
                                    'url': url,
                                    'content': text[:1000]  # First 1000 chars
                                })
                    except:
                        continue
            
            return content_data
            
        except Exception as e:
            logger.error(f"[Browser] Content retrieval error: {e}")
            return []
