backend/ai-automation-service/requirements.txt
# requirements.txt - Definitive list for Python 3.11 on Render

# --- Core Web Server & API ---
# These versions are known to have stable wheels for Python 3.11
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
numpy

# --- Key Integrations & AI ---
openai
PyGithub
web3

# --- Essential Utilities ---
requests
httpx
python-multipart
redis
sqlalchemy
psycopg2-binary

anthropic>=0.25.0
google-generativeai>=0.4.0
huggingface-hub>=0.20.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
radon>=6.0.0
pytest>=7.0.0